% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_extract_pred.R
\name{logit_extract_pred}
\alias{logit_extract_pred}
\title{Extract predictions from logistic GLM}
\usage{
logit_extract_pred(model_name, new_data = new_data)
}
\arguments{
\item{model_name}{Name of variable containing GLM}

\item{new_data}{Name of data frame containing new data to predict over}
}
\value{
A tibble with predictions and confidence intervals of model predictions
}
\description{
Extract predictions from logistic GLM
}
\examples{
# Simulate data for GLM
set.seed(2012)
gender <- sample(c(0,1), size = 100, replace = TRUE)
age <- round(runif(100, 18, 80))
xb <- -9 + 3.5*gender + 0.2*age
p <- 1/(1 + exp(-xb))
y <- rbinom(n = 100, size = 1, prob = p)
data_sim <- data.frame(y, gender, age)
data_sim <- data_sim \%>\%
  dplyr::mutate(y = as.factor(y),
                gender = as.factor(gender))

# Run model
mod1 <- glm(y ~ gender + age,
            data = data_sim,
            family = binomial(link = "logit"))

# New data to predict over
ndata <- expand.grid(
  gender = levels(data_sim$gender),
  age = seq(min(data_sim$age),
            max(data_sim$age),
            l = 100))

# Extract model predictions
mod_pred <- logit_extract_pred(model_name = mod1, new_data = ndata)
}
